feat: [ANDROAPP-5125] revert move TeiAttributesProvider
