+++ app/src/main/java/org/dhis2/usescases/teiDashboard/DashboardRepositoryImpl.kt
@@ -11,7 +11,6 @@ import org.dhis2.commons.prefs.Preference
 import org.dhis2.commons.prefs.PreferenceProvider
 import org.dhis2.commons.resources.MetadataIconProvider
 import org.dhis2.tracker.extensions.profilePicturePath
-import org.dhis2.tracker.relationships.data.TeiAttributesProvider
 import org.dhis2.ui.MetadataIconData
 import org.dhis2.utils.ValueUtils
 import org.hisp.dhis.android.core.D2
